User Input with Validation
This update fixes the ratio output to use integer parts instead of decimal values and adds UI validation for numeric input. The calculator now parses user-entered values before calling the ratio logic and handles invalid or empty inputs with clear feedback.

+++ src/calculator/Calculations.java
@@ -25,7 +25,7 @@ public class Calculations {
 			tempSec = tempFir % tempSec;
 			tempFir = tempNum;
 		}
-		results[0] = (firNum/tempFir)+":"+(secNum/tempFir);
+		results[0] = ((int)(firNum/tempFir))+":"+((int)(secNum/tempFir));
 		
 		// Simple ratio calculation, 1:x
 		results[1] = "1:"+String.format("%.3f", secNum/firNum);
